Knowledge umbauen

+++ blog/models.py
@@ -1,14 +1,14 @@
 from django.conf import settings
 from django.db import models
 from django.utils import timezone
-
+from django import forms
 
 class Post(models.Model):
     author = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     title = models.CharField(max_length=200)
     text = models.TextField()
     text2 = models.TextField()
-    knowledge = models.PositiveSmallIntegerField()
+    knowledge = models.IntegerField()
     created_date = models.DateTimeField(default=timezone.now)
     published_date = models.DateTimeField(blank=True, null=True)
